refactor(tier-refresh): lift matchSection into shared module
